# backend/user/api.yaml
openapi: 3.0.0
info:
  title: User service API
  description: API description in Markdown.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
paths:
...
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestRegistration'
            #Swagger does not generate example for oneOf
            example: '{
                      "auth_info": {
                        "email": "string",
                        "password": "string"
                      },
                      "personal_info": {
                        "first_name": "string",
                        "second_name": "string",
                        "middle_name": "string",
                        "date_of_birth": "2021-07-27"
                      },
                      "register_type": "PreUniversity",
                      "register_confirm": {
                        "registration_number": "string",
                        "one_time_password": "string"
                      }
                    }'
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseRegistration'
        '400':
          description: Bad request or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '409':
          description: Username already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /register/internal:
    post:
      summary: Register an internal user
      description: Register an internal user, admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestRegistrationInternal'
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseRegistrationInternal'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '409':
          description: Username already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /preregister:
    post:
      summary: Register an unconfirmed user with a one-time password
      description: Register an unconfirmed user with a one-time password, creators and admins only
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responsePreregister'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'

  # Authentication
  /login:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestLogin'
      security: []
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                type: string
                example: access_token_cookie=eyJ0eXAi...; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseLogin'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '401':
          description: Wrong credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /refresh:
    post:
      summary: Refresh JWT token for current user
      security:
        - JWTRefreshToken: [ ]
        - CSRFRefreshToken: [ ]
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                type: string
                example: access_token_cookie=eyJ0eXAi...; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseRefresh'
  /logout:
    post:
      summary: Logout current user
      security:
        - JWTAcessToken: []
        - CSRFAcessToken: [] #framework limitations
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              schema:
                description: Set both access_token_cookie and refresh_token_cookie
                type: string
                example: access_token_cookie=deleted; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT

  # User info
  /user/self:
    get:
      summary: Get common info for current user
      security:
        - JWTAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseUserSelf'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/{user_id}:
    get:
      summary: Get common info for a different user
      description: Get info about any user by its id, only for admins and creators
      security:
        - JWTAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseUserAdmin'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/all:
    get:
      summary: Get common info for all users
      description: Get info about all users, only for admins and creators
      security:
        - JWTAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseUserAll'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/by-group/{group_id}:
    get:
      summary: Get common info for different users
      description: Get info about any user by their group, only for admins and creators
      security:
        - JWTAcessToken: [ ]
      parameters:
        - in: path
          description: ID of the group
          name: group_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseUserByGroup'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'


  # Password
  /user/self/password:
    post:
      summary: Change password for current user
      security:
        - JWTAcessToken: []
        - CSRFAcessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestPasswordSelf'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/{user_id}/password:
    post:
      summary: Change password for another user
      description: Change password for another user, admins only
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestPasswordAdmin'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'

  # Permissions
  /user/{user_id}/role:
    put:
      summary: Set the role of any user
      description: Set the role of another user, only for admins
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestUserRole'
      responses:
        '200':
          description: OK
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/{user_id}/type:
    put:
      summary: Set the type of any user
      description: Set the type of another user, only for admins
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestUserType'
      responses:
        '200':
          description: OK
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '409':
          description: User is missing required info (e.g. university info for a university student)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'

  # Personal info
  /user/self/personal:
    get:
      summary: Get personal info for current user
      security:
        - JWTAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responsePersonalSelf'
        '404':
          description: Personal info is not set or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
  /user/{user_id}/personal:
    get:
      summary: Get personal info for a different user
      description: Get personal info about any user by its id, only for admins and creators
      security:
        - JWTAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responsePersonalAdminGet'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: Personal info is not set or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
    patch:
      summary: Set personal info for a user
      description: Set personal info about any user by its id, only for admins
      security:
        - JWTAcessToken: [ ]
        - CSRFAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the user
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/requestPersonalAdminPatch'
      responses:
        '200':
          description: OK
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '409':
          description: Personal info is not set and request is not full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'

  # University info
  /user/self/university:
    get:
      summary: Get university student info for current user
      security:
        - JWTAcessToken: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseUniversitySelf'
        '404':
          description: User not found or is not a university student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
...
  /group/{group_id}:
    get:
      summary: Get any group
      description: Get any group, only for admins and creators
      security:
        - JWTAcessToken: [ ]
      parameters:
        - in: path
          description: Id of the group
          name: group_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responseGroup'
        '403':
          description: Invalid role of current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
...
components:
  securitySchemes:

  ######################################################################################################################
  ##                                             Security components                                                  ##
  ######################################################################################################################
    
    JWTAcessToken:
      description: Authentication with JWT stored in cookies
      type: apiKey
      in: cookie
      name: access_token_cookie
    JWTRefreshToken:
      description: JWT stored in cookies to refresh an acess token
      type: apiKey
      in: cookie
      name: refresh_token_cookie
    CSRFAcessToken:
      description: CSRF acess token
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
    CSRFRefreshToken:
      description: CSRF refresh token for "remember me" functionality
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
...
    requestRegistration:
      title: 'Registration request info'
      oneOf:
        - $ref: '#/components/schemas/requestRegistrationSchool'
        - $ref: '#/components/schemas/requestRegistrationUniversity'
...
    responseRegistration:
      $ref: '#/components/schemas/typeUserInfo'

    # /register/internal
    requestRegistrationInternal:
      $ref: '#/components/schemas/typeAuthCredentials'
    responseRegistrationInternal:
      $ref: '#/components/schemas/typeUserInfo'

    # /preregister
    responsePreregister:
      $ref: '#/components/schemas/typePreregisterInfo'

    # /login
...
    responseLogin:
      $ref: '#/components/schemas/typeCSRFPair'

    # /refresh
    responseRefresh:
      $ref: '#/components/schemas/typeCSRFPair'

    # /logout

    # /user/self
...